refactor: break out code and tree
