+++ pkg/ast/tree.go
@@ -1,0 +1,9 @@
+package ast
+
+import (
+	"errors"
+)
+
+func (c *node) TreeString() (string, error) {
+	return "", errors.New("Not implemented!")
+}
